rename module compile -> compiler
