+++ src/compiler/compose.rs
@@ -152,3 +152,11 @@ healthcheck:
         )
     }
 }
+
+impl crate::compiler::Compile for Compose {
+    type Input = crate::type_check::BuildContext;
+
+    fn compile(input: Self::Input) -> Result<Self, String> {
+        todo!()
+    }
+}
